Add files via upload
Welcome to the Rock, Paper, Scissors game! This simple web-based game is implemented using HTML, CSS, and JavaScript. Test your luck and strategy by playing against the computer in this classic game.

How to Play:
Accessing the Game: To play the game, simply open the index.html file in a web browser that supports HTML5.

Game Rules: The rules of the game are simple:

Rock beats Scissors
Scissors beat Paper
Paper beats Rock
Making a Choice:   Click on one of the three buttons: "Rock," "Paper," or "Scissors" to make your choice.
Computer's Choice:   The computer's choice will be randomly generated.
Outcome:   The game will determine the winner based on the choices made by you and the computer. The result (win, lose, or draw) will be displayed on the screen.
Restart:   If you want to play again, click the "Play Again" button to reset the game and start a new round.

Files and Directory Structure
index.html: The main HTML file that displays the game interface and handles user interactions.
styles.css: The CSS file containing styles for the game interface.
script.js: The JavaScript file containing the game logic and interaction handling.
About the Implementation
HTML: The structure of the game is built using HTML. Buttons are used to represent the player's choices.
CSS: The game interface is styled using CSS to make it visually appealing and user-friendly.
JavaScript: The game logic is implemented using JavaScript. Event listeners capture user clicks, and random number generation is utilized for the computer's choice. The game outcome is determined by comparing the choices and displaying the result on the screen.